Modularize tirestino types
