+++ src/tirest/types/level.ts
@@ -1,0 +1,5 @@
+export interface Level {
+  number: number
+
+  linesToClear: number
+}
