+++ src/tirest/types/progress.ts
@@ -1,0 +1,9 @@
+export interface ProgressInfo {
+  clearedLines: number
+  score: number
+}
+
+export interface Progress {
+  totals: ProgressInfo
+  byLevel: ProgressInfo[]
+}
