+++ src/tirest/types/settings.ts
@@ -1,0 +1,12 @@
+export interface ColorPalette {
+  name: string
+  blocks: [string, string][]
+}
+
+export interface Settings {
+  musicVolume: number
+
+  inputDelay: number
+
+  selectedColorPalette: ColorPalette
+}
