+++ src/tirest/types/tirest.ts
@@ -1,0 +1,42 @@
+import type { UUID } from 'crypto'
+
+import type { Position } from './utils'
+import type { Level } from './level'
+import type { Progress } from './progress'
+import type { Settings } from './settings'
+
+export type GameState = 'Playing' | 'Paused' | 'GameOver'
+
+export interface Tirest {
+  score: number
+
+  fieldId: UUID
+  tirestinoQueueId: UUID
+
+  currentTirestinoId: number | null
+  heldTirestinoId: number | null
+
+  droppingFrom: Position
+
+  prevTime: DOMHighResTimeStamp
+  prevInputTime: DOMHighResTimeStamp
+  prevAutoFallTime: DOMHighResTimeStamp
+
+  hasManuallyDropped: boolean
+  hasRotated: boolean
+  hasHeld: boolean
+  hasEscaped: boolean
+  hasSelected: boolean
+
+  gameState: GameState
+
+  selectedMenuItem: number | null
+
+  combo: number
+
+  level: Level
+
+  progress: Progress
+
+  settings: Settings
+}
